httpd: Move authentication to its own module
